Unified generic output
